fix(conversations): probe each Ollama endpoint separately, not one for all stages
doctor/preflight each validated every conversations stage's Ollama model
against a single arbitrarily-chosen endpoint, so a model routed to a second
host (e.g. compact.extractive_backend on box.local while ask.backend stays
on localhost) was reported missing against a host that was never queried.
collect_backend_configs's dedup key was also (provider, model) only, which
silently dropped one endpoint when two stages shared a model name.

- Extend the dedup key to (provider, model, endpoint) and group Ollama
  backends by endpoint (group_ollama_backends_by_endpoint); probe each
  distinct endpoint once via a shared probe_ollama_tags helper, used by
  both doctor and preflight instead of two near-duplicated blocks.
- Add mur_common::config::DEFAULT_OLLAMA_ENDPOINT (pub) and point
  default_ollama_endpoint(), the two probe call sites, and
  config_migrate.rs's former private copy all at the single constant.
- Fix clippy::field_reassign_with_default in
  rewriter_does_not_fall_through_to_the_answer_stage_override (struct
  literal + ..Default::default()).
- Split conversations_cmd.rs (1775 lines after the above) into
  conversations_cmd/{mod,doctor,preflight}.rs so doctor/preflight own
  their command bodies while the shared probe helpers stay
  module-private in mod.rs, reachable from both via `super`.
- Add two pinning tests for the endpoint-grouping fix.
